Add files via upload
This is experimental code that can work either as a mobile decoder (for loco speed direction etc), or as static (accessory) decoder. 
The static stuff looks for switch changes and sends messages to rocrail allowing its use as a block sensor.
It also reacts to rocrail commands to allow it to move servos for points. 
This version can change SV's under rocrail control, but does not yet permit CV changes ..