Add files via upload
Updated, with switch / case selection replacing multiple if statements to increase processing speed
Servo Output selection is now based on "OUTPUT / PULSE"  selection in Configuration in I/O 
Setting CV[8]=8  will set defaults  (loco address 3 etc.) All io as inputs...

This version is for a Mobile (loco) decoder with an RFID reader.., but is essentially thesme as a static decoder.

TODO: make setup of IO fully dependant on the SV settings, so ports can be servo, inputs or LED drivers. 
(currently either inputs or servo("pulse") outputs.
Put "If Defines" in code to select if static or mobile decoder.